+++ Assets/UI/CrosshaireController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8cb25d52628db784da15a6a8b10ff952
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
